VM traffic stop after agent restart.
Closes-bug: #1794934
During an agent restart, the vif gets deleted. With server mode
configuration of vhostuser, we should not delete the socket
file in /var/run/vrouter.

+++ dpdk/vr_uvhost_client.c
@@ -8,6 +8,7 @@
 #include "vr_dpdk.h"
 #include "vr_uvhost_client.h"
 #include "vr_uvhost_util.h"
+#include "vr_uvhost_msg.h"
 
 static vr_uvh_client_t vr_uvh_clients[VR_UVH_MAX_CLIENTS];
 
@@ -62,7 +63,8 @@ vr_uvhost_del_client(vr_uvh_client_t *vru_cl)
     vr_uvhost_del_fds_by_arg(vru_cl);
 
     vru_cl->vruc_fd = -1;
-    unlink(vru_cl->vruc_path);
+    if (vru_cl->vruc_vhostuser_mode == VRNU_VIF_MODE_CLIENT)
+        unlink(vru_cl->vruc_path);
 
     return;
 }
